Move sortColumn to be with SortedTable
